Continue a optimization run
To do so, we store basically all important inforamtion (population, fitness values, bracket states, ..) as a pickle object.
By reloading this object, we can continue training from this state.
A thing to discuss is the current bracket state. I am not quite certain if there is another way to make sure that the brackets are correct. Also, this functionality is more about continuing and not warmstarting.
The difference for me between warmstarting and restarting is that for warmstarting we should think about reloading multiple states and thus combining the populations per budget might not be so simple.
However, restarting or continuing is more easy, because we just have to restore the old state. Happy to discuss this difference.

Also, I have added a short logger that writes the results similar to Hpbandsters BOHB to file. (TODO: config.json is still missing).